requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
geopy==2.4.1
plotly==5.18.0
tensorflow==2.16.1
xgboost==2.0.2
lightgbm==4.1.0
torch==2.2.1
transformers==4.35.2
aiohttp==3.9.1
asyncio==3.4.3
websockets==12.0
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.0
dash==2.14.1
dash-bootstrap-components==1.5.0
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
